+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_DESCRIPTION_BEGIN
+### NEXT_VERSION_DESCRIPTION_END
 ## [3.11.2](https://github.com/yoomoney-gradle-plugins/artifact-release-plugin/pull/17) (20-04-2021)
 
 * Снижено количество detekt нарушений 41 -> 18.