+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=PATCH
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Изменен DateTimeFormatter с dd-MM-YYYY на dd-MM-yyyy, т.к. YYYY отдает "week year", что некорректно.
 ### NEXT_VERSION_DESCRIPTION_END
 ## [5.0.0](https://github.com/yoomoney-gradle-plugins/artifact-release-plugin/pull/20) (21-06-2021)