+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-### NEXT_VERSION_TYPE=PATCH
-### NEXT_VERSION_DESCRIPTION_BEGIN
+## [5.0.1](https://github.com/yoomoney-gradle-plugins/artifact-release-plugin/pull/21) (21-07-2021)
+
 * Изменен DateTimeFormatter с dd-MM-YYYY на dd-MM-yyyy, т.к. YYYY отдает "week year", что некорректно.
-### NEXT_VERSION_DESCRIPTION_END
+
 ## [5.0.0](https://github.com/yoomoney-gradle-plugins/artifact-release-plugin/pull/20) (21-06-2021)
 
 * **breaking changes** artifact-release-plugin теперь коммитит только файлы